refactored messaging-service src/main packages structure
