--- messaging-service/src/main/java/com/krypton/messagingservice/Message.java
@@ -1,8 +1,0 @@
-package com.krypton.messagingservice;
-
-public class Message
-{
-    public String name;
-
-    public Message() {}
-}
