+++ messaging-service/src/main/java/com/krypton/messagingservice/MessagingServiceApplication.java
@@ -3,8 +3,10 @@ package com.krypton.messagingservice;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
+@EnableFeignClients
 @EnableEurekaClient
 public class MessagingServiceApplication {
