+++ messaging-service/src/main/java/com/krypton/messagingservice/config/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package com.krypton.messagingservice;
+package com.krypton.messagingservice.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
@@ -21,6 +21,6 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer
     @Override
     public void registerStompEndpoints(StompEndpointRegistry registry)
     {
-        registry.addEndpoint("/s/gs-guide-websocket").setAllowedOrigins("*").withSockJS();
+        registry.addEndpoint("/s/websocket-server").setAllowedOrigins("*").withSockJS();
     }
 }
